# src/main/java/br/com/socialconnect/api/beneficiarios/controller/BeneficiarioController.java
package br.com.socialconnect.api.beneficiarios.controller;

import br.com.socialconnect.api.beneficiarios.dto.BeneficiarioRequestDTO;
import br.com.socialconnect.api.beneficiarios.dto.BeneficiarioResponseDTO;
import br.com.socialconnect.api.beneficiarios.service.BeneficiarioService;
import jakarta.validation.Valid;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.net.URI;

@RestController
@RequestMapping("/api/v1/beneficiarios")
public class BeneficiarioController {

    private final BeneficiarioService service;

    public BeneficiarioController(BeneficiarioService service) {
        this.service = service;
    }

    // ✅ Chamando o método "listar" com os 3 parâmetros exatos
//    @GetMapping
//    public ResponseEntity<Page<BeneficiarioResponseDTO>> listar(
//            @RequestParam(required = false) String nome,
//            @RequestParam(required = false) String cpf,
//            @PageableDefault(size = 10, sort = "idBeneficiario", direction = Sort.Direction.ASC) Pageable pageable) {
//
//        return ResponseEntity.ok(service.listar(nome, cpf, pageable));
//    }
    @GetMapping
    public ResponseEntity<Page<BeneficiarioResponseDTO>> listar(
            @RequestParam(required = false) String nome,
            @RequestParam(required = false) String cpf,
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "10") int size,
            @RequestParam(defaultValue = "idBeneficiario,asc") String sort) {

        // Sanitiza o parâmetro sort (remove colchetes, aspas, espaços)
        String sortLimpo = sort.replaceAll("[\\[\\]\" ]", "");

        // Divide em campo e direção
        String[] sortParts = sortLimpo.split(",");
        String campo = sortParts[0];
        Sort.Direction direcao = sortParts.length > 1 && sortParts[1].equalsIgnoreCase("desc")
                ? Sort.Direction.DESC
                : Sort.Direction.ASC;

        // Cria o Pageable manualmente
        Pageable pageable = PageRequest.of(page, size, Sort.by(direcao, campo));

        return ResponseEntity.ok(service.listar(nome, cpf, pageable));
    }


    //    @GetMapping("/{idBeneficiario}")
//    public ResponseEntity<BeneficiarioResponseDTO> buscarPorId(@PathVariable Long idBeneficiario) {
//        //return ResponseEntity.ok(service.buscarPorId(idBeneficiario));
//        return service.buscarPorId(idBeneficiario) // ou repository.findById(id)
//                .map(ResponseEntity::ok)
//                .orElse(ResponseEntity.notFound().build());
//    }
@GetMapping("/{idBeneficiario}")
public ResponseEntity<BeneficiarioResponseDTO> buscarPorId(@PathVariable("idBeneficiario") Long idBeneficiario) {
    BeneficiarioResponseDTO response = service.buscarPorId(idBeneficiario);
    return ResponseEntity.ok(response);
}

    @PostMapping
    public ResponseEntity<BeneficiarioResponseDTO> criar(
            @Valid @RequestBody BeneficiarioRequestDTO dto) {
        BeneficiarioResponseDTO salvo = service.criar(dto);
        URI location = URI.create("/api/v1/beneficiarios/" + salvo.idBeneficiario());
        return ResponseEntity.created(location).body(salvo);
    }

    @DeleteMapping("/{idBeneficiario}")
    public ResponseEntity<Void> deletar(@PathVariable("idBeneficiario") Long idBeneficiario) {
        service.deletar(idBeneficiario);
        return ResponseEntity.noContent().build();
    }
}