# src/main/java/br/com/socialconnect/api/beneficiarios/dto/BeneficiarioPatchDTO.java
package br.com.socialconnect.api.beneficiarios.dto;

// DTO para PATCH: todos os campos opcionais
public record BeneficiarioPatchDTO(
        String nome,
        String telefone,
        String endereco,
        String situacaoVulnerabilidade
) {}