# src/main/java/br/com/socialconnect/api/doacoes/controller/DoacaoController.java
package br.com.socialconnect.api.doacoes.controller;

import br.com.socialconnect.api.doacoes.dto.DoacaoResponseDTO;
import br.com.socialconnect.api.doacoes.model.TipoDoacao;
import br.com.socialconnect.api.doacoes.service.DoacaoService;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import java.time.LocalDate;

@RestController
@RequestMapping("/api/v1/doacoes")
public class DoacaoController {

    private final DoacaoService service;

    public DoacaoController(DoacaoService service) {
        this.service = service;
    }

    @GetMapping
    public ResponseEntity<Page<DoacaoResponseDTO>> listar(
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dataInicio,
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dataFim,
            @RequestParam(required = false) TipoDoacao tipo,
            @PageableDefault(size = 10) Pageable pageable) { // ← Removido o sort daqui

        return ResponseEntity.ok(service.listar(dataInicio, dataFim, tipo, pageable));
    }
//    @GetMapping
//    public ResponseEntity<Page<DoacaoResponseDTO>> listar(
//            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dataInicio,
//            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dataFim,
//            @RequestParam(required = false) TipoDoacao tipo,
//            @PageableDefault(size = 10, sort = "dataDoacao", direction = Sort.Direction.DESC) Pageable pageable) {
//
//        return ResponseEntity.ok(service.listar(dataInicio, dataFim, tipo, pageable));
//    }
}